requirements.txt
azure-storage-blob>=12.19.0
openai>=1.12.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0
python-dotenv>=1.0.0